Fix issue with proto randomly not showing.
I assume the original issue (#190) was caused by the order in which pages
are included & rendered. This presumably caused the global variable
not to be set soon enough and when re-rendered get loaded from a cache.

This change extends the `Haml::Helpers` module and removes the global variable.
Not sure where the best location is for this file is.